move SearchSidebar to search-ui
