+++ client/search-ui/src/index.ts
@@ -4,3 +4,6 @@ export * from './input/SearchBox'
 export * from './documentation/ModalVideo'
 export * from './results/StreamingSearchResultsList'
 export * from './results/progress/StreamingProgress'
+export * from './results/sidebar/revisions'
+export * from './results/sidebar/FilterLink'
+export * from './results/sidebar/SearchSidebar'
