+++ client/search-ui/src/results/sidebar/SearchReference.module.scss
