+++ client/search-ui/src/results/sidebar/SearchSidebar.module.scss
