+++ client/search-ui/src/results/sidebar/SearchSidebar.story.tsx
@@ -4,15 +4,13 @@ import React from 'react'
 // eslint-disable-next-line no-restricted-imports
 import create from 'zustand'
 
-import { BuildSearchQueryURLParameters, SearchQueryState } from '@sourcegraph/search'
+import { BrandedStory } from '@sourcegraph/branded/src/components/BrandedStory'
+import { BuildSearchQueryURLParameters, SearchPatternType, SearchQueryState } from '@sourcegraph/search'
 import { QuickLink, SearchScope } from '@sourcegraph/shared/src/schema/settings.schema'
 import { Filter } from '@sourcegraph/shared/src/search/stream'
 import { EMPTY_SETTINGS_CASCADE } from '@sourcegraph/shared/src/settings/settings'
 import { NOOP_TELEMETRY_SERVICE } from '@sourcegraph/shared/src/telemetry/telemetryService'
 import { buildSearchURLQuery } from '@sourcegraph/shared/src/util/url'
-
-import { WebStory } from '../../../components/WebStory'
-import { SearchPatternType } from '../../../graphql-operations'
 
 import { SearchSidebar, SearchSidebarProps } from './SearchSidebar'
 
@@ -143,10 +141,10 @@ const filters: Filter[] = [
     })),
 ]
 
-add('empty sidebar', () => <WebStory>{() => <SearchSidebar {...defaultProps} />}</WebStory>)
+add('empty sidebar', () => <BrandedStory>{() => <SearchSidebar {...defaultProps} />}</BrandedStory>)
 
 add('with everything', () => (
-    <WebStory>
+    <BrandedStory>
         {() => (
             <SearchSidebar
                 {...defaultProps}
@@ -154,5 +152,5 @@ add('with everything', () => (
                 filters={filters}
             />
         )}
-    </WebStory>
+    </BrandedStory>
 ))
