+++ client/search-ui/src/results/sidebar/SearchSidebarSection.module.scss
