+++ client/search-ui/src/results/sidebar/SearchSidebarSection.test.tsx
