+++ client/search-ui/src/results/sidebar/SearchSidebarSection.tsx
