+++ client/search-ui/src/results/sidebar/helpers.test.ts
