+++ client/search-ui/src/results/sidebar/helpers.ts
