+++ client/web/src/search/results/sidebar/Revisions.tsx
@@ -3,8 +3,8 @@ import classNames from 'classnames'
 import React from 'react'
 
 import { dataOrThrowErrors, gql } from '@sourcegraph/http-client'
-import { QueryUpdate } from '@sourcegraph/search'
-import { SyntaxHighlightedSearchQuery } from '@sourcegraph/search-ui'
+import { FilterLink, RevisionsProps, SyntaxHighlightedSearchQuery, TabIndex } from '@sourcegraph/search-ui'
+import styles from '@sourcegraph/search-ui/src/results/sidebar/SearchSidebarSection.module.scss'
 import { GitRefType } from '@sourcegraph/shared/src/schema'
 import { FilterType } from '@sourcegraph/shared/src/search/query/filters'
 import { useTemporarySetting } from '@sourcegraph/shared/src/settings/temporary/useTemporarySetting'
@@ -16,9 +16,6 @@ import {
     SearchSidebarGitRefsVariables,
     SearchSidebarGitRefFields,
 } from '../../../graphql-operations'
-
-import { FilterLink } from './FilterLink'
-import styles from './SearchSidebarSection.module.scss'
 
 const DEFAULT_FIRST = 10
 export const GIT_REVS_QUERY = gql`
@@ -140,21 +137,6 @@ const RevisionList: React.FunctionComponent<RevisionListProps> = ({
             ) : null}
         </>
     )
-}
-
-export enum TabIndex {
-    BRANCHES,
-    TAGS,
-}
-
-export interface RevisionsProps {
-    repoName: string
-    onFilterClick: (updates: QueryUpdate[]) => void
-    query: string
-    /**
-     * This property is only exposed for storybook tests.
-     */
-    _initialTab?: TabIndex
 }
 
 export const Revisions: React.FunctionComponent<RevisionsProps> = React.memo(
